uploading Authentication for all the Service classes of TOURIST

+++ Unreal_Journey_Engine/Unreal_Journey_Engine/Controllers/BookingController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Cors;
+using Unreal_Journey_Engine.AuthFilters;
 
 namespace Unreal_Journey_Engine.Controllers
 {
@@ -19,9 +20,23 @@ namespace Unreal_Journey_Engine.Controllers
         #region Get All Bookings
         [HttpGet]
         [Route("all")]
+        [Logged]
         public HttpResponseMessage Get_All_Bookings()
         {
+            var authorizationHeader = Request.Headers.Authorization?.ToString();
+            var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+            if (current_user_Type == "Admin")
+            {
 
+            }
+            else
+            {
+                var responseMessage = new
+                {
+                    Message = "You are not allowed to see all bookings. Only admin can"
+                };
+                return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
+            }
             var data = BookingService.Get();
             if (data.Count > 0)
             {
@@ -42,6 +57,7 @@ namespace Unreal_Journey_Engine.Controllers
         #region Get Single Booking
         [HttpGet]
         [Route("{id}")]
+        [Logged]
         public HttpResponseMessage Get(int id)
         {
             try
@@ -70,40 +86,54 @@ namespace Unreal_Journey_Engine.Controllers
         #region  Create
         [HttpPost]
         [Route("create")]
+        [Logged]
         public HttpResponseMessage Create_Booking(BookingDTO dto)
         {
             try
             {
-
-                if (dto != null)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Tourist" || current_user_Type == "Admin")
                 {
-                    var decision = BookingService.Create(dto);
-                    if (decision)
+                    if (dto != null)
                     {
-                        var responseMessage = new
+                        var decision = BookingService.Create(dto);
+                        if (decision)
                         {
-                            Message = "Booking Created"
-                        };
-                        return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                            var responseMessage = new
+                            {
+                                Message = "Booking Created"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                        }
+                        else
+                        {
+                            var responseMessage = new
+                            {
+                                Message = "Failed to Create Booking"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        }
+
                     }
                     else
                     {
                         var responseMessage = new
                         {
-                            Message = "Failed to Create Booking"
+                            Message = "Provide Booking Data to Create Booking"
                         };
-                        return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
                     }
-
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Provide Booking Data to Create Booking"
+                        Message = "You are not allowed to Create bookings. Only tourist and admin can"
                     };
-                    return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
@@ -115,40 +145,54 @@ namespace Unreal_Journey_Engine.Controllers
         #region Update
         [HttpPut]
         [Route("update")]
+        [Logged]
         public HttpResponseMessage Update_Booking_Info(BookingDTO dto)
         {
             try
             {
-
-                if (dto != null)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Tourist" || current_user_Type == "Admin")
                 {
-                    var decision = BookingService.Update(dto);
-                    if (decision)
+                    if (dto != null)
                     {
-                        var responseMessage = new
+                        var decision = BookingService.Update(dto);
+                        if (decision)
                         {
-                            Message = "Booking Updated"
-                        };
-                        return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                            var responseMessage = new
+                            {
+                                Message = "Booking Updated"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                        }
+                        else
+                        {
+                            var responseMessage = new
+                            {
+                                Message = "Failed to Update Booking"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        }
+
                     }
                     else
                     {
                         var responseMessage = new
                         {
-                            Message = "Failed to Update Booking"
+                            Message = "Provide Booking Data to Create Booking"
                         };
-                        return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
                     }
-
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Provide Booking Data to Create Booking"
+                        Message = "You are not allowed to Update bookings. Only tourist and admin can"
                     };
-                    return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
@@ -160,23 +204,38 @@ namespace Unreal_Journey_Engine.Controllers
         #region Delete
         [HttpDelete]
         [Route("delete/{id}")]
+        [Logged]
         public HttpResponseMessage Delete_Booking_Info(int id)
         {
             try
             {
-                var data = BookingService.Delete(id);
-                if (data)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Tourist" || current_user_Type == "Admin")
                 {
-                    return Request.CreateResponse(HttpStatusCode.OK, data);
+                    var data = BookingService.Delete(id);
+                    if (data)
+                    {
+                        return Request.CreateResponse(HttpStatusCode.OK, data);
+                    }
+                    else
+                    {
+                        var responseMessage = new
+                        {
+                            Message = "Booking Not Found"
+                        };
+                        return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                    }
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Booking Not Found"
+                        Message = "You are not allowed to delete bookings. Only tourist and admin can"
                     };
-                    return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
