+++ Unreal_Journey_Engine/Unreal_Journey_Engine/Controllers/Tour_PackageController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Cors;
+using Unreal_Journey_Engine.AuthFilters;
 
 namespace Unreal_Journey_Engine.Controllers
 {
@@ -18,6 +19,7 @@ namespace Unreal_Journey_Engine.Controllers
         #region Get All Tours
         [HttpGet]
         [Route("all")]
+        [Logged]
         public HttpResponseMessage Get_All_Tours()
         {
 
@@ -41,6 +43,7 @@ namespace Unreal_Journey_Engine.Controllers
         #region Get Single Tour
         [HttpGet]
         [Route("{id}")]
+        [Logged]
         public HttpResponseMessage Get(int id)
         {
             try
@@ -66,88 +69,117 @@ namespace Unreal_Journey_Engine.Controllers
         }
         #endregion Get Single Tour
 
-        #region Signup / Create
+        #region  Create
         [HttpPost]
         [Route("create")]
+        [Logged]
         public HttpResponseMessage Create_Tour_Package(Tour_PackageDTO dto)
         {
             try
             {
-
-                if (dto != null)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Admin")
                 {
-                    var decision = Tour_PackageService.Create(dto);
-                    if (decision)
+                    if (dto != null)
                     {
-                        var responseMessage = new
+                        var decision = Tour_PackageService.Create(dto);
+                        if (decision)
                         {
-                            Message = "Account Created"
-                        };
-                        return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                            var responseMessage = new
+                            {
+                                Message = "Tour Package Created"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                        }
+                        else
+                        {
+                            var responseMessage = new
+                            {
+                                Message = "Failed to Create Tour Package"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        }
+
                     }
                     else
                     {
                         var responseMessage = new
                         {
-                            Message = "Failed to Create Account"
+                            Message = "Provide Tour Data to Create Package"
                         };
-                        return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
                     }
-
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Provide Tour Data to Create Account"
+                        Message = "You are not allowed to create tour packages. Only admin can"
                     };
                     return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
                 return Request.CreateResponse(HttpStatusCode.InternalServerError, ex.Message);
             }
         }
-        #endregion Signup / Create
+        #endregion Create
 
         #region Update
         [HttpPut]
         [Route("update")]
+        [Logged]
         public HttpResponseMessage Update_Tour_Info(Tour_PackageDTO dto)
         {
             try
             {
 
-                if (dto != null)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Admin")
                 {
-                    var decision = Tour_PackageService.Update(dto);
-                    if (decision)
+                    if (dto != null)
                     {
-                        var responseMessage = new
+                        var decision = Tour_PackageService.Update(dto);
+                        if (decision)
                         {
-                            Message = "Account Update"
-                        };
-                        return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                            var responseMessage = new
+                            {
+                                Message = "Tour package Update"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                        }
+                        else
+                        {
+                            var responseMessage = new
+                            {
+                                Message = "Failed to Update Tour package"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        }
+
                     }
                     else
                     {
                         var responseMessage = new
                         {
-                            Message = "Failed to Update Account"
+                            Message = "Provide Tour Data to Create Account"
                         };
-                        return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
                     }
-
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Provide Tour Data to Create Account"
+                        Message = "You are not allowed to update tour package. Only Admin can"
                     };
-                    return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
@@ -159,23 +191,38 @@ namespace Unreal_Journey_Engine.Controllers
         #region Delete
         [HttpDelete]
         [Route("delete/{id}")]
+        [Logged]
         public HttpResponseMessage Delete_Tour_Info(int id)
         {
             try
             {
-                var data = Tour_PackageService.Delete(id);
-                if (data)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Admin")
                 {
-                    return Request.CreateResponse(HttpStatusCode.OK, data);
+                    var data = Tour_PackageService.Delete(id);
+                    if (data)
+                    {
+                        return Request.CreateResponse(HttpStatusCode.OK, data);
+                    }
+                    else
+                    {
+                        var responseMessage = new
+                        {
+                            Message = "Tour package Not Found"
+                        };
+                        return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                    }
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Account Not Found"
+                        Message = "You are not allowed to delete tour packagaes. Only admin can"
                     };
-                    return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
