+++ Unreal_Journey_Engine/Unreal_Journey_Engine/Controllers/Tour_ReviewController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Cors;
+using Unreal_Journey_Engine.AuthFilters;
 
 namespace Unreal_Journey_Engine.Controllers
 {
@@ -18,6 +19,7 @@ namespace Unreal_Journey_Engine.Controllers
         #region Get All Tour Reviews
         [HttpGet]
         [Route("all")]
+        [Logged]
         public HttpResponseMessage Get_All_Tour_Reviews()
         {
 
@@ -38,9 +40,10 @@ namespace Unreal_Journey_Engine.Controllers
         }
         #endregion Get All Tour Reviews
 
-        #region Get Single Tour
+        #region Get Single Tour Review
         [HttpGet]
         [Route("{id}")]
+        [Logged]
         public HttpResponseMessage Get(int id)
         {
             try
@@ -69,40 +72,54 @@ namespace Unreal_Journey_Engine.Controllers
         #region Post / Create
         [HttpPost]
         [Route("create")]
+        [Logged]
         public HttpResponseMessage Create_Tour_Review(Tour_ReviewDTO dto)
         {
             try
             {
-
-                if (dto != null)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Tourist")
                 {
-                    var decision = Tour_ReviewService.Create(dto);
-                    if (decision)
+                    if (dto != null)
                     {
-                        var responseMessage = new
+                        var decision = Tour_ReviewService.Create(dto);
+                        if (decision)
                         {
-                            Message = "Review Posted"
-                        };
-                        return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                            var responseMessage = new
+                            {
+                                Message = "Review Posted"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                        }
+                        else
+                        {
+                            var responseMessage = new
+                            {
+                                Message = "Failed to Post Review"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        }
+
                     }
                     else
                     {
                         var responseMessage = new
                         {
-                            Message = "Failed to Post Review"
+                            Message = "Provide Tour Review Data to Post a Review"
                         };
-                        return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
                     }
-
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Provide Tour Review Data to Post a Review"
+                        Message = "You are not allowed to give any review.  Only tourist can."
                     };
-                    return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
@@ -114,40 +131,54 @@ namespace Unreal_Journey_Engine.Controllers
         #region Update
         [HttpPut]
         [Route("update")]
+        [Logged]
         public HttpResponseMessage Update_Tour_Info(Tour_ReviewDTO dto)
         {
             try
             {
-
-                if (dto != null)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Tourist" || current_user_Type == "Admin")
                 {
-                    var decision = Tour_ReviewService.Update(dto);
-                    if (decision)
+                    if (dto != null)
                     {
-                        var responseMessage = new
+                        var decision = Tour_ReviewService.Update(dto);
+                        if (decision)
                         {
-                            Message = "Account Update"
-                        };
-                        return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                            var responseMessage = new
+                            {
+                                Message = "Review Updated"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.OK, responseMessage);
+                        }
+                        else
+                        {
+                            var responseMessage = new
+                            {
+                                Message = "Failed to Update Review"
+                            };
+                            return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        }
+
                     }
                     else
                     {
                         var responseMessage = new
                         {
-                            Message = "Failed to Update Account"
+                            Message = "Provide Tour Data to Update Review"
                         };
-                        return Request.CreateResponse(HttpStatusCode.NotAcceptable, responseMessage);
+                        return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
                     }
-
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Provide Tour Data to Create Account"
+                        Message = "You are not allow to update tour review.  Only tourist and admin can."
                     };
-                    return Request.CreateResponse(HttpStatusCode.PreconditionFailed, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
@@ -159,23 +190,38 @@ namespace Unreal_Journey_Engine.Controllers
         #region Delete
         [HttpDelete]
         [Route("delete/{id}")]
+        [Logged]
         public HttpResponseMessage Delete_Tour_Review(int id)
         {
             try
             {
-                var data = Tour_ReviewService.Delete(id);
-                if (data)
+                var authorizationHeader = Request.Headers.Authorization?.ToString();
+                var current_user_Type = User_Info_Provider.Get_User_Role(authorizationHeader);
+                if (current_user_Type == "Tourist" || current_user_Type == "Admin")
                 {
-                    return Request.CreateResponse(HttpStatusCode.OK, data);
+                    var data = Tour_ReviewService.Delete(id);
+                    if (data)
+                    {
+                        return Request.CreateResponse(HttpStatusCode.OK, data);
+                    }
+                    else
+                    {
+                        var responseMessage = new
+                        {
+                            Message = "Review Not Found"
+                        };
+                        return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                    }
                 }
                 else
                 {
                     var responseMessage = new
                     {
-                        Message = "Review Not Found"
+                        Message = "You are not allowed to delete review. Only tourist and admin can."
                     };
-                    return Request.CreateResponse(HttpStatusCode.NotFound, responseMessage);
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, responseMessage);
                 }
+                
             }
             catch (Exception ex)
             {
